+++ astra-sim/system/Roofline.hh
@@ -17,12 +17,12 @@ public:
   void set_bandwidth(double bandwidth);
   void set_neg_y_intercept(double neg_y_intercept);
   void set_peak_perf(double peak_perf);
-  uint64_t get_perf(double operational_intensity);
+  double get_perf(double operational_intensity);
 
 private:
   double bandwidth;
-  uint64_t neg_y_intercept;
-  uint64_t peak_perf;
+  double neg_y_intercept;
+  double peak_perf;
 };
 
 }
